% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{copy_from_dropbox}
\alias{copy_from_dropbox}
\title{Copy file from local dropbox directory}
\usage{
copy_from_dropbox(from_file, to_file)
}
\arguments{
\item{from_file}{Local Dropbox file to copy}

\item{to_file}{Output local file path}
}
\description{
Copy file from local dropbox directory
}

