% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{copy_to_dropbox}
\alias{copy_to_dropbox}
\title{Copy file to local dropbox directory}
\usage{
copy_to_dropbox(from_file, to_file)
}
\arguments{
\item{from_file}{Local file to copy to Dropbox}

\item{to_file}{Output Dropbox file path}
}
\description{
Copy file to local dropbox directory
}

