% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{gff2bed12}
\alias{gff2bed12}
\title{GFF3 to BED12 converter}
\usage{
gff2bed12(obj, file = NULL)
}
\arguments{
\item{obj}{GRanges object}

\item{file}{file path to GFF3 input file}
}
\value{
out BED12 formatted dataframe
}
\description{
Will convert a GFF3 GRanges object to a BED12 dataframe
}
\details{
Provide a GRanges object in GFF3 format or a file path to a GFF3 formatted
annotation file and return a dataframe that meets the specs of a BED12 file
as specified by UCSC
}

