% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{kill_nb}
\alias{kill_nb}
\title{Kill the Jekyll Serve}
\usage{
kill_nb()
}
\description{
This function will stop the jekyll server using a system call
}
\details{
{ \code{kill_nb} calls 'pkill -f jekyll' to kill the jekyll serve
    process. }
}

