% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{new_entry}
\alias{new_entry}
\title{Create a New Notebook Entry}
\usage{
new_entry(title = "New entry", dir = "_rmd", subdir = "new_entry",
  skeleton = "_build/skeleton_entry.Rmd")
}
\arguments{
\item{title}{The title of the entry}

\item{dir}{The directory all of your .Rmd entries are in}

\item{subdir}{The subdirectory with \code{dir} that you want your new entry
to be placed it}

\item{skeleton}{The filepath of a skeleton of a new entry}
}
\description{
Thanks to https://github.com/brendan-R/brocks for inspiring this function
}
\details{
This function will add a new entry within the '_rmd' directory under a
subdirectory by default and create a blank .R script to accompany it

{ \code{new_entry} will create a .R file, and a .Rmd file
  withinin a subdirectory, with names created by running \code{title} through
  \code{\link{filenamise}}. The .R file will contain a short note mentioning
  that it accompanies the .Rmd file, which will contain the same text as the
  file supplied by \code{skeleton} paramter. }
}

