% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{serve_nb}
\alias{serve_nb}
\title{Serve the Jekyll Notebook}
\usage{
serve_nb(host = "0.0.0.0", port = "4000", config = "_config_dev.yml")
}
\arguments{
\item{host}{The host address to serve the jekyll site on}

\item{port}{The port to make the jekyll site accessible on}

\item{config}{The config file to use to for jekyll}
}
\description{
This function will make a system call using jekyll serve on the specified
host address, port, and using the specified config file
}
\details{
{ \code{serve_nb} will build the notebook only using jekyll. It will
   not rebuild any .Rmd files. \code{\link{build_nb}} needs to be run in
   order to rebuild any changed .Rmd files. }
}

